requirements.txt
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cloudscraper==1.2.71
cssselect==1.3.0
datarecorder==3.6.2
downloadkit==2.0.7
drissionpage==4.1.0.18
et-xmlfile==2.0.0
filelock==3.18.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
lxml==5.4.0
naked==0.1.32
numpy==2.2.5
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycryptodome==3.22.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
schedule==1.2.2
shellescape==3.8.1
six==1.17.0
soupsieve==2.7
tldextract==5.3.0
tqdm==4.67.1
typing-extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
websocket-client==1.8.0
